backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic-settings
python-dotenv
passlib
bcrypt
python-multipart
chromadb
fastembed
neo4j
